+++ src/components/home/favorite/Favorite.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState } from "react";
 import { FavoriteButton } from "src/components/home/favorite/FavoriteButton";
 import { LikeButton } from "src/components/home/like/LikeButton";
 import { useFetchFavorits } from "src/components/home/favorite/useFetchFavorits";
-import { useFetchLikes } from "src/components/home/like/useFetchLikes";
+import { useFetchLikes } from "src/hooks/useFetchLikes";
 import { supabase } from "src/libs/supabase";
 import type { definitions } from "src/types/supabaseTypes";
 import type { PostgrestResponse } from "@supabase/postgrest-js";
