+++ src/components/home/post/PostList.tsx
@@ -5,10 +5,10 @@ import { AiOutlinePlus } from "react-icons/ai";
 import { LikeButton } from "src/components/home/like/LikeButton";
 import { FavoriteButton } from "src/components/home/favorite/FavoriteButton";
 import { useFetchFavorits } from "src/components/home/favorite/useFetchFavorits";
-import { useFetchLikes } from "src/components/home/like/useFetchLikes";
+import { useFetchLikes } from "src/hooks/useFetchLikes";
 import { EditPageLinkButton } from "../edit/EditPageLinkButton";
 import { CommentButton } from "../comment/CommentButton";
-import { useGetPost } from "./useGetPost";
+import { useGetPost } from "../../../hooks/useGetPost";
 
 export const PostList = () => {
   const { fetchPosts, posts } = useGetPost(); //postsは表示する記事
