+++ src/components/mypage/User.tsx
@@ -1,14 +1,14 @@
 import { useState, useEffect, useCallback } from "react";
 import { supabase } from "src/libs/supabase";
 import { PersonalSearch } from "src/components/mypage/PersonalSearch";
-import { useFetchProfiles } from "src/components/mypage/useFetchProfiles";
+import { useFetchProfiles } from "src/hooks/useFetchProfiles";
 import Image from "next/image";
 import toast from "react-hot-toast";
 import { FavoriteButton } from "src/components/home/favorite/FavoriteButton";
 import { LikeButton } from "src/components/home/like/LikeButton";
-import { useFetchLikes } from "src/components/home/like/useFetchLikes";
+import { useFetchLikes } from "src/hooks/useFetchLikes";
 import { useFetchFavorits } from "src/components/home/favorite/useFetchFavorits";
-import { useFetchPosts } from "../home/post/useFetchPosts";
+import { useFetchPosts } from "../../hooks/useFetchPosts";
 
 export const User = () => {
   const { profile, fetchProfiles } = useFetchProfiles();
